# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinanIes",
        "contact": {
            "email": "derex@outlook.com.br"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/v1/user/{user}/associated": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "associated user with a role",
                "description": "associated user with a role",
                "operationId": "7e0e0253ecafc0de2a38fe36ec500781",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "user identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role_id": {
                                        "description": "role id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/{user}/dissociate": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "dissociate user with a role",
                "description": "dissociate user with a role",
                "operationId": "13a88ce7e8b38e9b5cdbf1dbff123b2a",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "user identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role_id": {
                                        "description": "role id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            }
        }
    }
}